+++ src/example.spec.ts
@@ -1,5 +1,0 @@
-describe('Example test', () => {
-  test('it should work', () => {
-    expect(2 + 3).toEqual(5);
-  });
-});